requirements.txt
# GTK4 Python bindings
PyGObject>=3.42.0

# Cron management
python-crontab>=3.0.0

# Additional utilities
croniter>=1.4.1
